3/5일 스터디

+++ controller/CartController.js
@@ -2,8 +2,6 @@ const ensureAuthorization = require("../auth"); // 인증 모듈
 const jwt = require('jsonwebtoken');
 const conn = require('../mariadb'); // db 모듈
 const {StatusCodes} = require('http-status-codes'); // status code 모듈
-const dotenv = require('dotenv');
-dotenv.config();
 
 
 // 장바구니 담기
@@ -38,7 +36,7 @@ const addToCart = (req,res) => {
 
 // 장바구니 아이템 목록 조회
 const getCartItems = (req,res) => {
-        const {selected} = req.body;
+        const {selected} = req.body || {} ;
 
         let authorization = ensureAuthorization(req,res);
 
@@ -54,8 +52,14 @@ const getCartItems = (req,res) => {
             let sql = `SELECT cartItems.id, book_id, title, summary, quantity, price 
                         FROM cartItems LEFT JOIN books 
                         ON cartItems.book_id = books.id
-                        WHERE user_id=? AND cartItems.id IN (?);`;
-            let values = [authorization.id, selected]
+                        WHERE user_id=? `;
+            let values = [authorization.id];
+
+            if (selected){ // 주문서 작성 시 '선택한 장바구니 목록 조회'
+                sql += `AND cartItems.id IN (?)`;
+                values.push(selected);
+            } 
+            
             conn.query(sql,values,
                 (err, results) => {
                     if(err) {
@@ -64,27 +68,36 @@ const getCartItems = (req,res) => {
                     }
 
                     res.status(StatusCodes.OK).json(results);
-        })
+            })
         }
-        
-
-
 };
 
 // 장바구니 아이템 삭제
 const removeCartItem = (req,res) => {
         const cartItemId = req.params.id; // cartItemId
         
-        let sql = `DELETE FROM cartItems WHERE id = ? `;
-        conn.query(sql, cartItemId, 
-            (err, results) => {
-                if(err) {
-                    console.log(err);
-                    return res.status(StatusCodes.BAD_REQUEST).end(); // BAD REQUEST
-                }
+        let authorization = ensureAuthorization(req,res);
 
-                res.status(StatusCodes.OK).json(results);
-        })
+        if(authorization instanceof jwt.TokenExpiredError) {
+            return res.status(StatusCodes.UNAUTHORIZED).json({
+                "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+            });
+        } else if (authorization instanceof jwt.JsonWebTokenError){ 
+            return res.status(StatusCodes.BAD_REQUEST).json({
+                "message" : "잘못된 토큰입니다."
+            });
+        }else {
+                let sql = `DELETE FROM cartItems WHERE id = ? `;
+            conn.query(sql, cartItemId, 
+                (err, results) => {
+                    if(err) {
+                        console.log(err);
+                        return res.status(StatusCodes.BAD_REQUEST).end(); // BAD REQUEST
+                    }
+
+                    res.status(StatusCodes.OK).json(results);
+            })
+        }
 };
 
 module.exports = {addToCart,getCartItems,removeCartItem};
